Explicitly cast types from OS Stat_t structs.

+++ node_unix.go
@@ -23,10 +23,10 @@ func toStatT(i interface{}) (statT, bool) {
 	return nil, false
 }
 
-func (s statUnix) dev() uint64   { return s.Dev }
-func (s statUnix) ino() uint64   { return s.Ino }
-func (s statUnix) nlink() uint64 { return s.Nlink }
-func (s statUnix) uid() uint32   { return s.Uid }
-func (s statUnix) gid() uint32   { return s.Gid }
-func (s statUnix) rdev() uint64  { return s.Rdev }
-func (s statUnix) size() int64   { return s.Size }
+func (s statUnix) dev() uint64   { return uint64(s.Dev) }
+func (s statUnix) ino() uint64   { return uint64(s.Ino) }
+func (s statUnix) nlink() uint64 { return uint64(s.Nlink) }
+func (s statUnix) uid() uint32   { return uint64(s.Uid) }
+func (s statUnix) gid() uint32   { return uint64(s.Gid) }
+func (s statUnix) rdev() uint64  { return uint64(s.Rdev) }
+func (s statUnix) size() int64   { return int64(s.Size) }
